+++ analytics/app/__init__.py
@@ -1,4 +1,4 @@
-"""Analytics dashboard application package."""
+"""
+Analytics dashboard application package.
+"""
 from .dashboard import app, server
-
-__all__ = ['app', 'server']
